requirements.txt
langchain
langgraph
openai
streamlit
requests
python-dotenv
pytest